requirements.txt
certifi==2018.4.16
google-api-python-client==1.6.7
httplib2==0.11.3
oauth2client==4.1.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
PyDrive==1.3.1
pyparsing==2.2.0
python-dateutil==2.7.3
PyYAML==3.12
rsa==3.4.2
six==1.11.0
tqdm==4.23.3
uritemplate==3.0.0
wincertstore==0.2
